requirements.txt

requests
python-telegram-bot==13.15
mailpytm
flask
